// src/client/schema.d.ts
export interface components {
    schemas: {
// ...
        CooldownSchema: {
            /**
             * Total Seconds
             * @description The total seconds of the cooldown.
             */
            total_seconds: number;
            /**
             * Remaining Seconds
             * @description The remaining seconds of the cooldown.
             */
            remaining_seconds: number;
            /**
             * Started At
             * Format: date-time
             * @description The start of the cooldown.
             */
            started_at: string;
            /**
             * Expiration
             * Format: date-time
             * @description The expiration of the cooldown.
             */
            expiration: string;
            /** @description The reason of the cooldown. */
            reason: components["schemas"]["ActionType"];
        };
// ...
        ItemResponseSchema: {
            data: components["schemas"]["ItemSchema"];
        };
        /** ItemSchema */
        ItemSchema: {
            /**
             * Name
             * @description Item name.
             */
            name: string;
            /**
             * Code
             * @description Item code. This is the item's unique identifier (ID).
             */
            code: string;
            /**
             * Level
             * @description Item level.
             */
            level: number;
            /**
             * Type
             * @description Item type.
             */
            type: string;
            /**
             * Subtype
             * @description Item subtype.
             */
            subtype: string;
            /**
             * Description
             * @description Item description.
             */
            description: string;
            /**
             * Conditions
             * @description Item conditions. If applicable. Conditions for using or equipping the item.
             */
            conditions?: components["schemas"]["ConditionSchema"][];
            /**
             * Effects
             * @description List of object effects. For equipment, it will include item stats.
             */
            effects?: components["schemas"]["SimpleEffectSchema"][];
            /** @description Craft information. If applicable. */
            craft?: components["schemas"]["CraftSchema"];
            /**
             * Tradeable
             * @description Item tradeable status. A non-tradeable item cannot be exchanged or sold.
             */
            tradeable: boolean;
            /**
             * Recyclable
             * @description Item recyclable status. A recyclable item can be recycled at the matching workshop.
             * @default false
             */
            recyclable: boolean;
        };
// ...
        MapSchema: {
            /**
             * Map Id
             * @description ID of the map.
             */
            map_id: number;
            /**
             * Name
             * @description Name of the map.
             */
            name: string;
            /**
             * Skin
             * @description Skin of the map.
             */
            skin: string;
            /**
             * X
             * @description Position X of the map.
             */
            x: number;
            /**
             * Y
             * @description Position Y of the map.
             */
            y: number;
            /** @description Layer of the map. */
            layer: components["schemas"]["MapLayer"];
            /** @description Access information for the map */
            access: components["schemas"]["AccessSchema"];
            /** @description Interactions available on this map. */
            interactions: components["schemas"]["InteractionSchema"];
        };
// ...
        ResourceSchema: {
            /**
             * Name
             * @description The name of the resource
             */
            name: string;
            /**
             * Code
             * @description The code of the resource. This is the resource's unique identifier (ID).
             */
            code: string;
            /**
             * Skill code
             * @description The skill required to gather this resource.
             */
            skill: components["schemas"]["GatheringSkill"];
            /**
             * Level
             * @description The skill level required to gather this resource.
             */
            level: number;
            /**
             * Drops
             * @description The drops of this resource.
             */
            drops: components["schemas"]["DropRateSchema"][];
        };
// ...
        StaticDataPage_MapSchema_: {
            /** Data */
            data: components["schemas"]["MapSchema"][];
            /** Total */
            total: number;
            /** Page */
            page: number;
            /** Size */
            size: number;
            /** Pages */
            pages: number;
        };
// ...
        StaticDataPage_ResourceSchema_: {
            /** Data */
            data: components["schemas"]["ResourceSchema"][];
            /** Total */
            total: number;
            /** Page */
            page: number;
            /** Size */
            size: number;
            /** Pages */
            pages: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}